+++ src/lektra.hpp
@@ -170,6 +170,7 @@ public:
     void TabsCloseOthers() noexcept;
     void Reshow_jump_marker() noexcept;
     void Show_recent_files_picker() noexcept;
+    void Copy_page_image() noexcept;
 
 #ifdef ENABLE_LLM_SUPPORT
     void ToggleLLMWidget() noexcept;